docs: csf 2 to 3
