+++ packages/core/src/TagsInput/TagsInput.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
 
 import { HvTagsInput } from ".";
-import { ControlledTagArray } from "./TagsInput.stories";
+import { ControlledTagArray } from "./stories/ControlledTagArray";
 
 describe("TagsInput examples", () => {
   describe("<ControlledTagArray />", () => {
